requirements.txt
anthropic>=0.40.0
supabase>=2.5.0
pydantic>=2.5.0
python-docx>=1.1.0
openpyxl>=3.1.2
voyageai>=0.2.3
firecrawl-py>=1.0.0
twilio>=9.0.0
googlemaps>=4.10.0
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
loguru>=0.7.2
tenacity>=8.2.3
pytest>=8.0.0
black>=24.0.0
ruff>=0.3.0
